git svn: revert default behavior for --minimize-url
This reverts the --minimize-url behavior change that
appeared recently in commit 80f8654ce0dbe71c82187e669cb8ce701d906de3
("Fix branch detection when repository root is inaccessible").

However, we now allow the option to be turned off by allowing
"--no-minimize-url" so people with limited-access setups can
still take advantage of the fix in
80f8654ce0dbe71c82187e669cb8ce701d906de3.

Also document the behavior and default settings of minimize-url
in the manpage for the first time.

This introduces a temporary UI regression to allow t9141 to pass
that will be reverted (fixed) in the next commit.

Signed-off-by: Eric Wong <normalperson@yhbt.net>